no-mistakes(review): Move capture helper to subdirectory to stop startup load error
