+++ pi/.pi/agent/extensions/no-mistakes-pane/capture.ts
